Design, Docs: 가이드 모달 토글 컨텐트 스타일 추가 및 수정, 가이드 데이터 수정

+++ asset/js/guideData.js
@@ -48,7 +48,7 @@ const guideData = [
     arguments: [
       { param: 'min', type: 'Float', detail: '최솟값' },
       { param: 'max', type: 'Float', detail: '최댓값' },
-      { param: 'round', type: 'Float', detail: '반올림 자릿수' },
+      { param: 'round', type: 'Integer', detail: '반올림 자릿수' },
     ],
     returns: 'Float',
   },
